Renaming of Default World plugin, added previous worlds.
