--- runelite-client/src/main/java/net/runelite/client/plugins/chatpreviousworld/PreviousWorldChatConfig.java
@@ -1,65 +1,0 @@
-/*
- * Copyright (c) 2018, PandahRS <https://github.com/PandahRS>
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package net.runelite.client.plugins.chatpreviousworld;
-
-import net.runelite.client.config.Config;
-import net.runelite.client.config.ConfigGroup;
-import net.runelite.client.config.ConfigItem;
-
-@ConfigGroup(
-	keyName = "chatpreviousworld",
-	name = "Previous worlds in chat",
-	description = "Displays previous worlds of the player in the chat after hopping or logging in."
-)
-public interface PreviousWorldChatConfig extends Config
-{
-	@ConfigItem(
-		keyName = "maxWorlds",
-		name = "Max world amount",
-		description = "Maximum amount of worlds to save."
-	)
-	default int getMaxWorldCount()
-	{
-		return 5;
-	}
-
-	@ConfigItem(
-		keyName = "pastworlds",
-		name = "",
-		description = "",
-		hidden = true
-	)
-	default String pastWorldList()
-	{
-		return "";
-	}
-
-	@ConfigItem(
-		keyName = "pastworlds",
-		name = "",
-		description = ""
-	)
-	void pastWorldList(String worldList);
-}
